Update twitch-json.js
now there is a array of online and offline, and with a function, we are able to print out the array or objects. And also the functions gets called at the end of loading up the JSON data, so there is no longer a problem of undefined caused by it being asynchronous. (We might have to look into this to use this asynchronous data, so that its visually live)