Solved issue, cleaned some code
Added a temporary solution to where sometimes the display all wasn't displaying all initially because the JSON was still loading up the array of objects. A pause of 1 second did the trick.

Cleaned some console.log() code and comments

Added a function to display the array of objects in a loop. So far made it so that we see online/offline and channel name.